Sat, Jan 28, 2023, 5:33 PM -05:00

+++ kochan/10_character_strings/exercise_10/07_string_to_float.c
@@ -49,7 +49,7 @@ void get_digits(char arr[])
  */
 float string_to_float(char arr[])
 {
-    float converted = 0.0, set_decimal_point = 1.0;
+    double converted = 0.0, set_decimal_point = 1.0;
     char zero = '0';
     _Bool is_negative = 0;
     _Bool found_digit = 0;
